Finished the development of the ALU sub controllers for the arithemtic operations and the logical flow directed by the main ALU control unit.
Added notes of progress made throughout the development and update of the ALU during this section of the project.  Furthermore, the third Latex report template is implemented and is expected to be followed and completed, with revisions on the way.

This commit here is to mark the first draft of the report.